moved not found page into new folder
